rewards: more bit packing
just because uint96 is more common and still can hold a huge ETH balance

+++ packages/nouns-contracts/contracts/client-incentives/INounsClientTokenTypes.sol
@@ -20,11 +20,11 @@ interface INounsClientTokenTypes {
         /// @notice Whether the DAO has approved the client to withdraw their rewards.
         bool approved;
         /// @notice The amount of reward tokens this client has been rewarded.
-        uint104 rewarded;
+        uint96 rewarded;
         /// @notice The amount of tokens this client has withdrawn.
-        uint104 withdrawn;
+        uint96 withdrawn;
         /// @dev A gap for future storage needs.
-        uint40 __gap;
+        uint56 __gap;
         /// @notice The client's display name.
         string name;
         /// @notice The client's description, e.g. its URL.
